requirements.txt
pandas
gspread 
oauth2client
streamlit
python-dotenv